Refactoring

+++ src/hbs/factory/DoubleRoom.java
@@ -17,7 +17,7 @@ public class DoubleRoom implements IRoom {
 	}
 
 	@Override
-	public double getOccupancy() {
+	public int getOccupancy() {
 		return occupancy;
 	}
